+++ mne/tests/test_transforms.py
@@ -17,7 +17,7 @@ from mne.transforms import (invert_transform, _get_trans,
                             get_ras_to_neuromag_trans, _pol_to_cart,
                             quat_to_rot, rot_to_quat, _angle_between_quats,
                             _find_vector_rotation, _sph_to_cart, _cart_to_sph,
-                            _topo_to_sph, SphericalHarmonicTPSWarp)
+                            _topo_to_sph, SphericalSplineWarp)
 
 warnings.simplefilter('always')  # enable b/c these tests throw warnings
 
@@ -45,7 +45,7 @@ def test_tps():
     destination *= 2
     destination[:, 0] += 1
     # fit with 100 points
-    warp = SphericalHarmonicTPSWarp().fit(source[::2], destination[::2])
+    warp = SphericalSplineWarp().fit(source[::2], destination[::2])
     destination_est = warp.transform(source)
     assert_allclose(destination_est, destination, atol=1e-2)