+++ doc/whats_new.rst
@@ -17,7 +17,7 @@ Changelog
 
     - Add filter plotting functions :func:`mne.viz.plot_filter` and :func:`mne.viz.plot_ideal_filter` as well as filter creation function :func:`mne.filter.create_filter` by `Eric Larson`_
 
-    - Add TPS warping with spherical harmonic surface approximations in :class:`mne.transforms.SphericalHarmonicTPSWarp` by `Eric Larson`_
+    - Add 3D thin-plate spline warping with spherical harmonic surface approximations in :class:`mne.transforms.SphericalSplineWarp` by `Eric Larson`_
 
 BUG
 ~~~